+++ docs/releases.rst
@@ -16,7 +16,8 @@ Graphite-API releases
   database-backed finders such as Cyanite because it allows fetching all time
   series at once instead of sequentially.
 
-* Add ``minimumBelow`` and ``changed`` functions.
+* Add ``multiplySeriesWithWildcards``, ``minimumBelow`` and ``changed``
+  functions.
 
 * Add ``/metrics`` API call as an alias to ``/metrics/find``.